Add insert logic for dogs table
